fix(build): MQTTnet 4.3.x regression + 6 pre-existing compile errors
The previous commit assumed MQTTnet 5.x had absorbed IManagedMqttClient
into core. Wrong: the entire MQTTnet.Extensions.ManagedClient package
was REMOVED in 5.0.0 as a breaking change (see
https://github.com/dotnet/MQTTnet/wiki/Upgrading-guide). Pin both
packages to 4.3.* (floating) and restore the using directives that
the previous fix deleted.

Once MQTTnet resolved, dotnet publish revealed 6 more pre-existing
bugs that the previous build never caught:

- ingestion/Program.cs: missing `using IoTIngestion.Payloads;` for
  TelemetryEvent. Pass Serilog.ILogger to MEL-typed constructors via
  a LoggerFactory with .AddSerilog() adapter.
- ingestion/Metrics/MetricsServer.cs and analytics/Metrics/MetricsServer.cs:
  `Metrics.CreateCounter` resolved to the enclosing namespace
  (IoTIngestion.Metrics / IoTAnalytics.Metrics) instead of
  Prometheus.Metrics. Use fully qualified Prometheus.Metrics.
- analytics/Brokers/MqttConsumer.cs: MqttQualityOfServiceLevel lives
  in MQTTnet.Protocol, not MQTTnet.Client. Add `using MQTTnet.Protocol;`.
  IManagedMqttClient.SubscribeAsync takes IEnumerable<MqttTopicFilter>,
  not a single one — wrap in an array.
- ingestion/Brokers/MqttPublisher.cs: same MQTTnet.Protocol using.
- ingestion/Replay/Scheduler.cs: `devices[emitted % devices.Count]`
  — emitted is long, list indexer is int. Explicit cast.

Verified: dotnet publish -c Release succeeds for both services with
0 warnings, 0 errors. Storage (TypeScript) was also re-verified with
npx tsc -p tsconfig.json.

+++ services/ingestion/Replay/Scheduler.cs
@@ -99,7 +99,8 @@ public sealed class Scheduler
         {
             var row = rows[rowIdx % rows.Count];
             rowIdx++;
-            var device = devices[emitted % devices.Count];
+            // emitted je long; IReadOnlyList indexer prima int.
+            var device = devices[(int)(emitted % devices.Count)];
 
             var payload = BuildPayload(row, device);
             await publishAsync(payload).ConfigureAwait(false);
